backend/requirements.txt
aiohttp==3.9.3
fastapi==0.110.2
openai==1.23.2
pydantic==2.7.0
python-dotenv==1.0.1
Requests==2.31.0
uvicorn==0.29.0
